+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Changed `defn` to automatically export functions as public and added `defn-` for private declarations
 - Add support for `noalias` function parameter modifier
+- Add support for `nosuspend` expression/block keyword
 
 ## 0.1.0 (2020-12-20)